sample_scenarios/requirements.txt
amazon.ion==0.5.0
boto3==1.9.237
botocore==1.12.237
pyqldb>=1.0.0-rc.1